Put panelapp functions on dedicated modules
