+++ scout/commands/update/panelapp.py
@@ -5,7 +5,7 @@ import logging
 import click
 from flask.cli import current_app, with_appcontext
 
-from scout.load.panel import load_panelapp_green_panel
+from scout.load.panelapp import load_panelapp_green_panel
 from scout.server.extensions import store
 
 LOG = logging.getLogger(__name__)
